abstracted state selectors into their own module
