+++ src/js/reducers/index.js
@@ -6,7 +6,7 @@ import loginReducer from './loginReducer';
 import recentProjectsReducer from './recentProjectsReducer';
 import importOnlineReducer from './importOnlineReducer';
 import groupMenuReducer from './groupMenuReducer';
-import settingsReducer, * as fromSettingsReducer from './settingsReducer';
+import settingsReducer from './settingsReducer';
 import loaderReducer from './loaderReducer';
 import popoverReducer from './popoverReducer';
 import resourcesReducer from './resourcesReducer';
@@ -29,7 +29,7 @@ import missingVersesReducer from './missingVersesReducer';
 import wordAlignmentReducer from './wordAlignmentReducer';
 import localImportReducer from './localImportReducer';
 import { localeReducer as locale } from 'react-localize-redux';
-import localeSettings, * as fromLocaleSettings from './localeSettingsReducer';
+import localeSettings from './localeSettings';
 // combining reducers
 const rootReducers = combineReducers({
   locale,
@@ -65,48 +65,3 @@ const rootReducers = combineReducers({
 });
 
 export default rootReducers;
-
-// state selectors
-
-/**
- * Retrieves an application setting
- * @param state
- * @param key
- * @return {*}
- */
-export const getSetting = (state, key) =>
-  fromSettingsReducer.getSetting(state.settingsReducer, key);
-
-/**
- * Returns a list of loaded languages available for the app locale.
- * This is a wrapper around react-localize-redux
- * @param state
- * @return {Language[]}
- */
-export const getLocaleLanguages = (state) =>
-  fromLocaleSettings.getLanguages(state);
-
-/**
- * Returns the currently active app locale.
- * This is a wrapper around react-localize-redux
- * @param state
- * @return {Language}
- */
-export const getActiveLocaleLanguage = (state) =>
-  fromLocaleSettings.getActiveLanguage(state);
-
-/**
- * Checks if the locale has finished loading
- * @param state
- * @return {*}
- */
-export const getLocaleLoaded = (state) =>
-  fromLocaleSettings.getLocaleLoaded(state.localeSettings);
-
-/**
- * Checks if the locale settings screen is open
- * @param state
- * @return {*}
- */
-export const getLocaleSettingsOpen = (state) =>
-  fromLocaleSettings.getLocaleSettingsOpen(state.localeSettings);
