+++ src/js/reducers/localeSettings.js
